[BEAM-77] Move bigtable in IO
